requirements.txt
Pillow
Flask
firebase-admin
google-generativeai
google-cloud-secret-manager
flask-cors
bs4
duckduckgo-search
google-cloud-logging